fix: 약 후기 - 수정할 부분 수정

+++ src/main/java/com/ll/projectLimC/domain/medicine/medicine/dto/MedicineResponseDTO.java
@@ -46,7 +46,7 @@ public class MedicineResponseDTO {
     private double averageRating;
 
     @Schema(description = "해당 약품 전체 후기 횟수")
-    private long totalReviewCount;
+    private double totalReviewCount;
 
     @Schema(description = "약품 북마크 횟수")
     private long bookmarkCount;
